requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pandas>=2.1.3
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.0.325
langchain-google-genai>=0.0.5
pydantic>=2.5.1
